+++ pytorch_lightning/trainer/trainer.py
@@ -279,6 +279,7 @@ class Trainer(
             truncated_bptt_steps: Truncated back prop breaks performs backprop every k steps of
 
             resume_from_checkpoint: To resume training from a specific checkpoint pass in the path here.
+                This can be a URL.
 
             profiler:  To profile individual steps during training and assist in